requirements.txt
tqdm
torch
torchvision
tensorboard
s3fs
matplotlib
seaborn
numpy
pandas
scikit-learn
Pillow==10.1.0
scipy
python-dotenv
